Bugfix: scenario name alignment with rfill should ignore tags
When measuring the columnb width, only the last line of the scenario
name should be considered. Otherwise, the line number information is
insufficiently spaced (because it's missing the spacing chars
incorrectly subtracted due to the length of tags, etc. in the lines
Bugfix: scenario name alignment with rfill should ignore tags

When measuring the columnb width, only the last line of the scenario
name should be considered. Otherwise, the line number information is
insufficiently spaced (because it's missing the spacing chars
incorrectly subtracted due to the length of tags, etc. in the lines
above the scenario name.